Move feedback and discord information to ponies member sheets, make feedback on auto lootlogs, @ people when their speed times are broken
